feat: Only show articles up till current date
I write posts for the upcoming days and weeks and not on the exact date.
This change will onlly list the current articles and not future ones
the content will still be thre but hidden and only shown once the site
is built for the current day. I will need to look into how to create a
cronjob for doing this daily build, hopefull Travis CI will be able to help